Projeção com Perspectiva Clippando arquivos obj 3D

+++ CG_Sistema2D/src/modelo/TipoClipador.java
@@ -6,6 +6,6 @@ public interface TipoClipador {
 
 	void clipar(Graphics g, boolean clipping);
 
-	int[][] cliparPoligono(Graphics g, boolean clipping);
+	double[][] cliparPoligono(Graphics g, boolean clipping);
 
 }
